refactor(sqldb): share read-only SQL validation and JSON param coercion
- Move ValidateReadOnlyWarehouseSQL + CoerceJSONParamValue to pkg/sqldb
- Fix WITH/AS parsing (s[2:] after AS, word boundary on WITH/SELECT)
- BigQuery: keep translatePlaceholders only in connector; use shared validation
- Snowflake: same validation + coerce params before QueryContext
